+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## 1.2.0 - 2020-05-24
 
 ### Added
 - Added support for Verbb Wishlist